Move localized-strings closer to localization service
